Refactor EpochLookup module
